move more code to share between async and blocking
