+++ src/blocking/mod.rs
@@ -6,6 +6,4 @@ pub mod json;
 pub mod proto_decoder;
 #[cfg(feature = "proto_raw")]
 pub mod proto_raw;
-#[cfg(any(feature = "proto_decoder", feature = "proto_raw"))]
-mod proto_resolver;
 pub mod schema_registry;
