+++ src/blocking/proto_decoder.rs
@@ -1,13 +1,15 @@
-use crate::blocking::proto_resolver::{to_index_and_data, MessageResolver};
+use std::collections::hash_map::{Entry, RandomState};
+use std::collections::HashMap;
+
+use bytes::Bytes;
+use protofish::{Context, MessageValue, Value};
+
 use crate::blocking::schema_registry::{
     get_bytes_result, get_referenced_schema, get_schema_by_id_and_type, SrSettings,
 };
 use crate::error::SRCError;
+use crate::proto_resolver::{to_index_and_data, MessageResolver};
 use crate::schema_registry_common::{BytesResult, RegisteredSchema, SchemaType};
-use bytes::Bytes;
-use protofish::{Context, MessageValue, Value};
-use std::collections::hash_map::{Entry, RandomState};
-use std::collections::HashMap;
 
 #[derive(Debug)]
 pub struct ProtoDecoder {
@@ -130,10 +132,11 @@ fn to_resolve_context(
 
 #[cfg(test)]
 mod tests {
-    use crate::blocking::proto_decoder::ProtoDecoder;
-    use crate::blocking::schema_registry::SrSettings;
     use mockito::{mock, server_address};
     use protofish::Value;
+
+    use crate::blocking::proto_decoder::ProtoDecoder;
+    use crate::blocking::schema_registry::SrSettings;
 
     fn get_proto_hb_schema() -> &'static str {
         r#"syntax = \"proto3\";package nl.openweb.data;message Heartbeat {uint64 beat = 1;}"#
